statue function updated

+++ contracts/TGVItemShop.sol
@@ -4,14 +4,13 @@ import "./TGVConfig.sol";
 
 contract TGVItemShop is TGVConfig {
     
-    ////// 장비 DB
-    function getEquip(uint unit_num)public view returns (Equip)
+    function getEquip(uint8 unit_num) public view returns (Equip)
     {
         //unit_num - 석상 인덱스
         Equip memory equip = equipList[msg.sender][unit_num];
         return equip;
     }
-    function reinforceEquip(uint32 unit_num, uint32 equip_num) public
+    function reinforceEquip(uint8 unit_num, uint8 equip_num) public
     {
         if(equip_num==1)
             equipList[msg.sender][unit_num].hpEquipLevel += 1;
